requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Optimization (optional)
cvxpy>=1.3.0

# Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0

# Live trading (optional)
fastapi>=0.95.0
uvicorn>=0.21.0
websockets>=11.0
psycopg2-binary>=2.9.0
influxdb-client>=1.36.0
redis>=4.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
pyyaml>=6.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
hypothesis>=6.70.0